+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Bearer token insertion in FastAPI UI.
 
+### Fixed
+- Limit query size.
+- Add OBI copyright.
+
 ## [0.0.8] - 10.12.2024
 
 ### Fixed